fix bg
