+++ apps/website-sv/src/routes/tags/[tag]/+page.server.ts
@@ -1,5 +1,13 @@
 import { PostService } from "$lib/posts-service.server";
-import type { PageServerLoad } from "./$types";
+import type { EntryGenerator, PageServerLoad } from "./$types";
+
+export const entries: EntryGenerator = async () => {
+	const posts = await PostService.list();
+	const tags = posts.flatMap((post) => post.tags);
+	const unique = new Set(tags);
+
+	return Array.from(unique).map((tag) => ({ tag }));
+};
 
 export const load: PageServerLoad = async ({ params }) => {
 	const posts = await PostService.list();
