+++ apps/website-sv/src/routes/writing/+layout.ts
@@ -1,0 +1,1 @@
+export const prerender = true;
